Added showing the image provided by server as marker's icon

+++ app/src/main/java/com/pouyaheydari/sample/map/android/pojo/VehiclesData.kt
@@ -7,7 +7,7 @@ data class VehiclesData(val vehicles: List<Vehicle>)
 
 @JsonClass(generateAdapter = true)
 data class Vehicle(
-    val bearing: Int,
+    val bearing: Float,
     val image_url: String,
     val lat: Double,
     val lng: Double,
